struct_method_moved_to_trait_must_use_added
